+++ components/src/core/Drawer/DrawerNavGroup.tsx
@@ -127,12 +127,14 @@ export const DrawerNavGroup: React.FC<DrawerNavGroupProps> = (props) => {
     const {
         // Inheritable Props
         /* eslint-disable @typescript-eslint/no-unused-vars */
+        activeChevronColor,
         activeItemBackgroundColor,
         activeItemBackgroundShape,
         activeItemFontColor,
         activeItemIconColor,
         backgroundColor,
         chevron,
+        chevronColor,
         collapseIcon,
         disableActiveItemParentStyles,
         divider,