+++ components/src/core/Drawer/types.tsx
@@ -10,6 +10,9 @@ export type DrawerContextProps = {
 };
 
 export type SharedStyleProps = {
+    /** chevron color for the active item */
+    activeChevronColor?: string;
+
     /** Background color for the active item */
     activeItemBackgroundColor?: string;
 
@@ -25,6 +28,9 @@ export type SharedStyleProps = {
 
     /** Color used for the background of the element */
     backgroundColor?: string;
+
+    /** The color used for the chevron icon */
+    chevronColor?: string;
 
     /**
      * Whether to show a line between all items
@@ -102,10 +108,12 @@ export type NavItemSharedStyleProps = {
 export type AllSharedProps = SharedStyleProps & NavItemSharedStyleProps;
 
 export const SharedStylePropTypes = {
+    activeChevronColor: PropTypes.string,
     activeItemBackgroundColor: PropTypes.string,
     activeItemFontColor: PropTypes.string,
     activeItemIconColor: PropTypes.string,
     backgroundColor: PropTypes.string,
+    chevronColor: PropTypes.string,
     divider: PropTypes.bool,
     itemFontColor: PropTypes.string,
     itemIconColor: PropTypes.string,